key state refactor
