--- web/src/containers/ApdKeyPerson/index.js
@@ -1,4 +1,0 @@
-import ApdKeyPersonForm from './ApdKeyPersonForm';
-import ApdKeyPersonReview from './ApdKeyPersonReview';
-
-export { ApdKeyPersonForm, ApdKeyPersonReview };
