+++ web/src/pages/apd/key-state-personnel/ApdStateProfileMedicaidOffice.js
@@ -13,11 +13,11 @@ import {
   setMedicaidOfficeCity,
   setMedicaidOfficeState,
   setMedicaidOfficeZip
-} from '../actions/editApd';
-import { t } from '../i18n';
-import { selectStateProfile } from '../reducers/apd.selectors';
-import { selectState } from '../reducers/user';
-import { STATES } from '../util';
+} from '../../../actions/editApd';
+import { t } from '../../../i18n';
+import { selectStateProfile } from '../../../reducers/apd.selectors';
+import { selectState } from '../../../reducers/user';
+import { STATES } from '../../../util';
 
 const dirTRoot = 'apd.stateProfile.directorAndAddress.director';
 const offTRoot = 'apd.stateProfile.directorAndAddress.address';
@@ -36,9 +36,11 @@ const ApdStateProfile = ({
 }) => {
   const { medicaidDirector, medicaidOffice } = stateProfile;
 
-  const handleChange = action => ({ target: { value } }) => {
-    action(value);
-  };
+  const handleChange =
+    action =>
+    ({ target: { value } }) => {
+      action(value);
+    };
 
   return (
     <Fragment>
