+++ web/src/pages/apd/key-state-personnel/KeyStatePersonnelReadOnly.js
@@ -1,7 +1,7 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 
-import Dollars from '../../components/Dollars';
+import Dollars from '../../../components/Dollars';
 
 const ApdStateProfile = ({ stateProfile, keyPersonnel }) => {
   const costByYear = person =>
